ml-processor/requirements.txt
kafka-python==2.0.2
redis==5.0.1
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
pytz==2023.3
